// src/interfaces/nfl/weeklyAdvStatsDef.d.ts
import { RawStatData } from './stats';
  
export interface RawWeeklyAdvStatDefData extends RawStatData, AdvDefData {}

export interface AdvDefData {
  player_weekly_id?: number,
  targets: number | null,
  completions_allowed: number | null,
  completion_pct: number | null,
  yards_allowed: number | null,
  yards_allowed_per_cmp: number | null,
  yards_allowed_per_tgt: number | null,
  rec_td_allowed: number | null,
  passer_rating_allowed: number | null,
  adot: number | null,
  air_yards_completed: number | null,
  yards_after_catch: number | null,
  blitzed: number | null,
  hurried: number | null,
  pressures: number | null,
  tackles_combined: number | null,
  tackles_missed: number | null,
  tackle_missed_pct: number | null,
}